HFP-3366 Fix dragging and resize uncategorized bug
The bug: When draggin a textItem from uncategorized, the textItems gets
position: fixed, making uncategorized smaller. If the textItem is then
let go so that it goes back into uncategorized, and the Move-to-Category
button is pressed very quickly after, the previousHeight state for
uncategorized has the possibility of being set to the height before the
textItem was released. This causes the maxHeight of the container to be
too low, forcing some textItems out into a third column almost
off-screen.

The solution: It turns out the current implementation of
resizeUncategorized( ) makes it possible to calculate and set maxHeight
before the dropdown has time to open. This means that previousHeight
is redundant and can be removed entirely. As a bonus, this makes the
code a lot cleaner!